requirements.txt
Flask==2.2.5
pandas
scikit-learn
gunicorn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
oauth2client